Fixed bug with scaling texts(see description)
This is hopefully one of the last submissions to the installer-part of this project, it is close to done. I fixed the scaling-issue with text by making them incredibly wide, if there are problems with text in the future, this probably did it. I reduced the available languages to english and german, french just wasnt worth it for an installer. I added a checkbox on the last window where a Readme.txt can be launched from.