Add latest, unpolished version of the main gui
This version still has some problems and unfinished features. The confirm-changes pop-up is not parented and doesnt have language support and the label "IfChangesReversed" does not work properly when used with "DropDownEmptyFolders". Furthermore, "factory settings" is unfinished. There probably also needs to be a rework of the button concepts. I would like to add the ability to add as many dir-references for the script to work in as i want to.