Add UpDownTemplate.ahk
This is a barebones version of the updown that is later supposed to be implemented into the main gui, but there are still some tests and some debugging left to do.
The function "FormatNumber( )" is now done and should format the parameter in form of string correctly. If the content of the edit, that is passed to the function, contains no numbers at all, the return will be "0". I am not sure yet, if I want to leave it at that or if I want it to throw an error instead. 
I would also like the contents of the edit to be saved in a more strict way timewise, as they are currently saved once the user unfocuses the edit or clicks the save button. This leaves huge room for bugs and unexpected behaviour.
The whole thing obviously needs to be polished up further to fit inside one single function in the best case.