Fixed function "FormatNumber( )" and added multiple general purpose functions described by their names.
"FormatNumber( )" now works as expected and transforms a string into a valid, positive number or throws an error if the string cannot be transformed. At the bottom of the function's code, there are two functions that are getting called, which remove the negative and positive signs that a number can have, "+" and "-". Removing the calls will keep the sign attached to the number. I changed the Label-name "hey" to the more fitting name of "UpdateContent". in said lable, there still remains an issue that consists of an infinite loop that is created when the entered value is not greater than the max or smaller than then min, in short, when the value is valid.
The code also needs to be further cleaned up and simplified.