Added new functions and fixed the infinite loop
To fix the infinite loop, I added multiple new functions. "IsInactive( )" is used to save the value of the edit after a set period of time, instead of waiting for the user to approve the input. The function checks for inactivity on base of the changing content of the edit. If the content doesnt change for the set period of time, the label in charge of saving the content, is called. This new, additive function fixes an earlier problem, but also introduces a new bug. The option to approve the input by hitting the enter-key is no longer available. I hope this is fixable as well.