Fixed the issue with the enter button not doing what it should
The previously limited options to submit a value are now back to the normal options of hitting enter, unfocusing the edit or letting enough time pass. To accomplish this, I had to reconstruct "IsActive( )" entirely, it works now with a stopwatch system, instead of a sleep system. In the next and hopefuly final version, I will have to fix a bug, where the number is formatted correctly by removing invalid characters, but is not updated in the edit. Other than that, I hope to finish this component soon.