Add new version with PopUpPromt( ) fully working and implemented
I have gone through a lot of iterations since the last upload of the main script. PopUpPromt( ) has been updated several times and is now implemted. I now have an easy and uniform way of creating a promt.
I have also added the necessary functions and labels for the UpDown, but it still needs to be connected to everything and polished.
I am thinking of adding some simple templates for PopUpPromt( ), so that calling a simple scematic of the promt can be called easier.
In the next uploaded iteration I will update the promt for unsaved changes when gui is closed. It still needs the third button for canceling the closing of settings.