Fixed issue with third button in PopUpPromt( )
PopUpPromt( ) is now completely finished and implemented, replacing the previously messy label system.
I might remove the labels as g-label from some buttons completely and put PopUpPromt as a direct function at their place. This would fight the huge ammount of labels that are currently in place.
I am still having issues with the UpDown, as the updown itself as a control seems to cause a bug, where the settings dont open at all and where the settings-button is practicly not working. I am going to look into this next and hopefully fix the issue.