Add fully working version, ready for implementation
I fixed the last few remaining issues. This version is theoreticly ready to be implemented, however, I want to make "WhileFocused:" and "WhileNotFocused:" a subroutine using settimer, instead of leaving them as normal labels. When used as labels, they lock the thread in them permanently, as they run until they call eachother, making them an infinite loop. The issue is that the code below won't continue.
I also added comments to the most important parts of the code.