Added a few functions and made small fixes to the existing ones
I added multiple functions to help improve and partly fix "FortmatNumber( )". 
One of the most notable changes in this version is the function "SetUpDown( )", that can set an UpDown to an edit later on to fix the reliability of the UpDown connecting to a "buddy"(an edit). I also removed the unnecessary SetTimer lines as I figured out that UpDownTemplate works just fine on the main Gui without SetTimer.