Big changes to the UpDownTemplate part of the main Gui; should be fully functional now
I made some big changes to the functionality of UpDownTemplate and its labels. I decided to not use SetTimer for the FocusCheck-part of UpDownTemplate, as I found it working fine without SetTimer. All functions of UpDownTemplate should be up and running fine now.
IMPORTANT: UpDownTemplate is still not included anywhere in the config, original variables and defaultsettings. You could say that its co-existing to the rest of the controls at this point in time, I will include it into the system of the other controls in future versions.

Apart from changes to UpDownTemplate, I also found that I still need to add a ReloadPromt( ) with the button to the config-file. I want it to tell the user clearly to be cautious about accessing and manipulationg the config by letting the user only continue to the config after he checks a checkbox with a clear warning as its text.
I also need to add language support to the ShowPromt.