Fixed all of the small bugs with "UpDownTemplate"
There were two little bugs:
1. the edit would'nt update in some cases after the content was formated. 
2. the edit would remain in a weird state and only partly accessible after the content was left blank.
Both of the issues are now fixed.
I also implemented the great majority of the internal code for "IfChangesReversed()", which should now include the new updown. 
I still have to write and read the values from and to the config at startup or shutdown.
The logic of the PopUpPromt when accessing the config has been implemented  but need further fine-tuning with the text.
There is no support for dynamic language with PopUpPromt() yet.